More code and tests

+++ contracts/core/IKODAV3Minter.sol
@@ -1,12 +1,13 @@
 // SPDX-License-Identifier: MIT
+
 pragma solidity 0.7.6;
 
 interface IKODAV3Minter {
 
     function mintToken(address _to, string calldata _uri) external returns (uint256 _tokenId);
 
-    function mintBatchEdition(uint256 _editionSize, address _to, string calldata _uri) external returns (uint256 _editionId);
+    function mintBatchEdition(uint96 _editionSize, address _to, string calldata _uri) external returns (uint256 _editionId);
 
-    function mintConsecutiveBatchEdition(uint256 _editionSize, address _to, string calldata _uri) external returns (uint256 _editionId);
+    function mintConsecutiveBatchEdition(uint96 _editionSize, address _to, string calldata _uri) external returns (uint256 _editionId);
 
 }
